Ensure headers are supplied as a map.
Previously, headers were accepted as any kind of data, but used as a map. This led to problems if the user put a vector in the EDN file instead of a map. The resulting error was obscure and did not include the offending form in the source.

This commit improves the situation very slightly by catching the error at parse time instead of at request time. We still need a more general error-catching and reporting approach, though.