Revised query interceptor syntax
Instead of splitting out 'find', 'where', and 'in' as keywords in the grammar we now have a single keyword 'q' that is followed by a Datalog query in exactly the format Datomic docs show. This not only makes the parsing easier (which is nice but not essential) but it makes the translation from Datomic docs to Vase descriptor a copy-and-paste thing instead of requiring any mental mapping from the user.