Parser accepts Http block with options
The http block is something new for Fern. It allows the user to specify all the options that would normally go in the service map, but directly from the config file. This reduces the need to have a "project" when all you want is to use Vase with a descriptor file.