Time to step up to a new parser design
We want to allow "modular syntax". That is, new libraries should be able to add new types of blocks to Fern. The current parser is very explicit. There are grammar rules for each type of top-level construct and adding a new one means modifying the Instaparse grammar. This was good enough to get started but needs to change now.

The new Instaparse grammar will be more abstract. It will mostly handle chunking a file into blocks. Each block will be handled by its own processor dispatched via a multimethod.

New block types can be added by adding methods on the multimethod.

This commit also formalizes error reporting from the processor. Errors come back as "markers" that indicate what the problem is and where it originated in the input.